feat: Home view prefetching part II
