+++ src/app/Scenes/HomeView/Sections/HomeViewSectionSales.tsx
@@ -73,9 +73,8 @@ export const HomeViewSectionSales: React.FC<HomeViewSectionSalesProps> = ({
 
   return (
     <Flex {...flexProps}>
-      <Flex px={2}>
-        <SectionTitle href={href} title={section.component?.title} onPress={onHeaderPress} />
-      </Flex>
+      <SectionTitle px={2} href={href} title={section.component?.title} onPress={onHeaderPress} />
+
       <CardRailFlatList
         prefetchUrlExtractor={(item) => item?.href}
         prefetchVariablesExtractor={(item) => ({ saleSlug: item?.slug })}
